+++ CodeChat_Server/CodeChat_Server/gen_py/CodeChat_Services/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['ttypes', 'constants', 'EditorPlugin', 'CodeChatClient']
+__all__ = ['ttypes', 'constants', 'EditorPlugin']