+++ CodeChat_Server/CodeChat_Server/gen_py/CodeChat_Services/ttypes.py
@@ -34,27 +34,6 @@ class CodeChatClientLocation(object):
     }
 
 
-class GetResultType(object):
-    html = 0
-    build = 1
-    errors = 2
-    command = 3
-
-    _VALUES_TO_NAMES = {
-        0: "html",
-        1: "build",
-        2: "errors",
-        3: "command",
-    }
-
-    _NAMES_TO_VALUES = {
-        "html": 0,
-        "build": 1,
-        "errors": 2,
-        "command": 3,
-    }
-
-
 class RenderClientReturn(object):
     """
     Attributes:
@@ -132,86 +111,12 @@ class RenderClientReturn(object):
 
     def __ne__(self, other):
         return not (self == other)
-
-
-class GetResultReturn(object):
-    """
-    Attributes:
-     - get_result_type
-     - text
-
-    """
-
-
-    def __init__(self, get_result_type=None, text=None,):
-        self.get_result_type = get_result_type
-        self.text = text
-
-    def read(self, iprot):
-        if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
-            iprot._fast_decode(self, iprot, [self.__class__, self.thrift_spec])
-            return
-        iprot.readStructBegin()
-        while True:
-            (fname, ftype, fid) = iprot.readFieldBegin()
-            if ftype == TType.STOP:
-                break
-            if fid == 1:
-                if ftype == TType.I32:
-                    self.get_result_type = iprot.readI32()
-                else:
-                    iprot.skip(ftype)
-            elif fid == 2:
-                if ftype == TType.STRING:
-                    self.text = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
-            else:
-                iprot.skip(ftype)
-            iprot.readFieldEnd()
-        iprot.readStructEnd()
-
-    def write(self, oprot):
-        if oprot._fast_encode is not None and self.thrift_spec is not None:
-            oprot.trans.write(oprot._fast_encode(self, [self.__class__, self.thrift_spec]))
-            return
-        oprot.writeStructBegin('GetResultReturn')
-        if self.get_result_type is not None:
-            oprot.writeFieldBegin('get_result_type', TType.I32, 1)
-            oprot.writeI32(self.get_result_type)
-            oprot.writeFieldEnd()
-        if self.text is not None:
-            oprot.writeFieldBegin('text', TType.STRING, 2)
-            oprot.writeString(self.text.encode('utf-8') if sys.version_info[0] == 2 else self.text)
-            oprot.writeFieldEnd()
-        oprot.writeFieldStop()
-        oprot.writeStructEnd()
-
-    def validate(self):
-        return
-
-    def __repr__(self):
-        L = ['%s=%r' % (key, value)
-             for key, value in self.__dict__.items()]
-        return '%s(%s)' % (self.__class__.__name__, ', '.join(L))
-
-    def __eq__(self, other):
-        return isinstance(other, self.__class__) and self.__dict__ == other.__dict__
-
-    def __ne__(self, other):
-        return not (self == other)
 all_structs.append(RenderClientReturn)
 RenderClientReturn.thrift_spec = (
     None,  # 0
     (1, TType.STRING, 'html', 'UTF8', None, ),  # 1
     (2, TType.I32, 'id', None, None, ),  # 2
     (3, TType.STRING, 'error', 'UTF8', None, ),  # 3
-)
-all_structs.append(GetResultReturn)
-GetResultReturn.thrift_spec = (
-    None,  # 0
-    (1, TType.I32, 'get_result_type', None, None, ),  # 1
-    (2, TType.STRING, 'text', 'UTF8', None, ),  # 2
 )
 fix_spec(all_structs)
 del all_structs